extracted StatsService functionality to StatsServiceLogic class, to prepare for being reused
